+++ example/lib/pages/sheet_page.dart
@@ -1,18 +1,24 @@
 import 'package:adaptive_dialog/adaptive_dialog.dart';
-import 'package:example/router.dart';
+import 'package:example/router/router.dart';
 import 'package:example/util/util.dart';
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:go_router/go_router.dart';
+
+class SheetRoute extends GoRouteData {
+  const SheetRoute();
+  @override
+  Widget build(BuildContext context) => const SheetPage();
+}
 
 class SheetPage extends StatelessWidget {
   const SheetPage({Key? key}) : super(key: key);
-
-  static const routeName = '/sheet';
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text(pascalCaseFromRouteName(SheetPage.routeName)),
+        title: Text(pascalCaseFromRouteName(GoRouter.of(context).location)),
       ),
       body: ListView(
         children: [
@@ -80,6 +86,34 @@ class SheetPage extends StatelessWidget {
                     key: 'helloKey',
                   ),
                 ],
+              );
+              logger.info(result);
+            },
+          ),
+          ListTile(
+            title: const Text('Theme builder'),
+            onTap: () async {
+              final result = await showModalActionSheet<String>(
+                context: context,
+                actions: [
+                  const SheetAction(
+                    icon: Icons.info,
+                    label: 'Hello',
+                    key: 'helloKey',
+                  ),
+                ],
+                builder: (context, child) => Theme(
+                  data: ThemeData(
+                    cupertinoOverrideTheme: const CupertinoThemeData(
+                      primaryColor: Colors.orange,
+                    ),
+                    listTileTheme: const ListTileThemeData(
+                      textColor: Colors.orange,
+                      iconColor: Colors.orange,
+                    ),
+                  ),
+                  child: child,
+                ),
               );
               logger.info(result);
             },