+++ lib/adaptive_dialog.dart
@@ -1,5 +1,7 @@
+export 'src/adaptive_dialog.dart';
 export 'src/adaptive_style.dart';
 export 'src/alert_dialog/alert_dialog.dart';
 export 'src/modal_action_sheet/modal_action_sheet.dart';
-export 'src/text_input_dialog/text_answer_dialog.dart';
-export 'src/text_input_dialog/text_input_dialog.dart';
+export 'src/text_input_dialog/show_text_answer_dialog.dart';
+export 'src/text_input_dialog/show_text_input_dialog.dart';
+export 'src/typedefs.dart';