+++ lib/src/alert_dialog/alert_dialog_action.dart
@@ -2,6 +2,8 @@ import 'package:adaptive_dialog/src/action_callback.dart';
 import 'package:adaptive_dialog/src/modal_action_sheet/sheet_action.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:intersperse/intersperse.dart';
+import 'package:macos_ui/macos_ui.dart';
 
 /// Used for specifying showAlertDialog's actions.
 @immutable
@@ -30,7 +32,7 @@ class AlertDialogAction<T> {
 }
 
 extension AlertDialogActionEx<T> on AlertDialogAction<T> {
-  Widget convertToCupertinoDialogAction({
+  Widget convertToIOSDialogAction({
     required ActionCallback<T> onPressed,
   }) {
     return CupertinoDialogAction(
@@ -39,6 +41,24 @@ extension AlertDialogActionEx<T> on AlertDialogAction<T> {
       textStyle: textStyle,
       onPressed: () => onPressed(key),
       child: Text(label),
+    );
+  }
+
+  Widget convertToMacOSDialogAction({
+    required ActionCallback<T> onPressed,
+  }) {
+    return PushButton(
+      buttonSize: ButtonSize.large,
+      isSecondary: isDestructiveAction || !isDefaultAction,
+      onPressed: () => onPressed(key),
+      child: Text(
+        label,
+        style: isDestructiveAction
+            ? const TextStyle(
+                color: CupertinoColors.destructiveRed,
+              )
+            : null,
+      ),
     );
   }
 
@@ -60,32 +80,48 @@ extension AlertDialogActionEx<T> on AlertDialogAction<T> {
 }
 
 extension AlertDialogActionListEx<T> on List<AlertDialogAction<T>> {
-  List<Widget> convertToCupertinoDialogActions({
+  List<Widget> convertToIOSDialogActions({
     required ActionCallback<T> onPressed,
   }) =>
-      map((a) => a.convertToCupertinoDialogAction(
-            onPressed: onPressed,
-          )).toList();
+      map(
+        (a) => a.convertToIOSDialogAction(
+          onPressed: onPressed,
+        ),
+      ).toList();
+
+  List<Widget> convertToMacOSDialogActions({
+    required ActionCallback<T> onPressed,
+    required ColorScheme colorScheme,
+  }) =>
+      map(
+        (a) => a.convertToMacOSDialogAction(
+          onPressed: onPressed,
+        ),
+      ).intersperse(const SizedBox(height: 8)).toList().reversed.toList();
 
   List<Widget> convertToMaterialDialogActions({
     required ActionCallback<T> onPressed,
     required Color destructiveColor,
     required bool fullyCapitalized,
   }) =>
-      map((a) => a.convertToMaterialDialogAction(
-            onPressed: onPressed,
-            destructiveColor: destructiveColor,
-            fullyCapitalized: fullyCapitalized,
-          )).toList();
+      map(
+        (a) => a.convertToMaterialDialogAction(
+          onPressed: onPressed,
+          destructiveColor: destructiveColor,
+          fullyCapitalized: fullyCapitalized,
+        ),
+      ).toList();
 
   List<SheetAction<T>> convertToSheetActions() =>
       where((a) => a.key != OkCancelResult.cancel)
-          .map((a) => SheetAction(
-                key: a.key,
-                label: a.label,
-                isDefaultAction: a.isDefaultAction,
-                isDestructiveAction: a.isDestructiveAction,
-              ))
+          .map(
+            (a) => SheetAction(
+              key: a.key,
+              label: a.label,
+              isDefaultAction: a.isDefaultAction,
+              isDestructiveAction: a.isDestructiveAction,
+            ),
+          )
           .toList();
 
   String? findCancelLabel() {