+++ lib/src/modal_action_sheet/sheet_action.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/widgets.dart';
-import 'package:meta/meta.dart';
 
 /// Used for specifying showModalActionSheet's actions.
 @immutable