+++ lib/src/typedefs.dart
@@ -1,0 +1,6 @@
+import 'package:flutter/widgets.dart';
+
+typedef AdaptiveDialogBuilder = Widget Function(
+  BuildContext context,
+  Widget child,
+);